requirements.txt
langchain
pyepsilla
sentence_transformers
streamlit
pypdf
python-dotenv
watchdog
openai
tiktoken
psycopg2-binary
langchain_experimental
chromadb